move pkg/proxy/server to cmd/kube-proxy/app
